Get a quick test working for endpoint and xml.
Add (hopefully) proper dependencies and exclusions for the endpoints
client project.

Add an async task to call the endpoint and populate the list view.

Add an async task to parse alert xml in the details view.

Tear out the dummy data classes that the wizard made for me.

Fix some xml deserialization errors (apparently to this 'simple'
framework, empty == missing).  Need to look into this to see if there's
some better way.